% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcoxt.R
\name{boxcoxt}
\alias{boxcoxt}
\title{Calculate the (one-parameter) Box-Cox transformation of a vector (given lambda)}
\usage{
boxcoxt(y, lambda = 0.1)
}
\arguments{
\item{x}{The vector to be transformed (Box-Cox transformation).}
}
\value{
A vector that is the Box-Cox transformation of \code{x}.
}
\description{
Calculate the (one-parameter) Box-Cox transformation of a vector (given lambda)
}
\details{
This is the one-parameter Box-Cox transformation.
This function does not calculate the optimal lambda value.
Using boxcox() in the MASS package is recommended for optimal lambda.
The default value of lambda is 0.1
All the elements of the input vector should be positive for Box-Cox transformation.
}
\examples{
boxcoxt(1:10)

}
