% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four.R
\name{four}
\alias{four}
\title{Calculate input to the fourth power (vector)}
\usage{
four(x, plot_it = FALSE)
}
\arguments{
\item{x}{The vector to be multiplied by its cube.}
}
\value{
A vector that is the fourth power of \code{x}.
}
\description{
That's it -- this function just makes a vector to the fourth power.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{four} function is quite simple.
     \item There's nothing else to say about \code{four}.
}
}
\examples{
four(1:10)
four(-5)
}
