% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{reciprocal}
\alias{reciprocal}
\title{Calculate the inverse (reciprocal) of a vector}
\usage{
reciprocal(x, plot_it = FALSE, na.omit = FALSE)
}
\arguments{
\item{x}{The vector to be inversed.}
}
\value{
A vector that is the reciprocal of \code{x}.
}
\description{
That's it -- this function just makes the inverse of a vector.
}
\details{
This function isn't complicated.

And it almost certainly doesn't need two paragraphs in the "Details"
section!

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{reciprocal} function is quite simple.
     \item There's nothing else to say about \code{reciprocal}.
}
}
\examples{
reciprocal(1:10)
reciprocal(-5)
}
